# codegen/openapi.yaml
openapi: 3.1.0
info:
  title: RPA Control-Plane API
  version: "1.0"
  description: >
    제어평면 REST API (v1). 운영 콘솔/외부 클라이언트가 호출하는 제어평면 엔드포인트의
    경로·메서드·요청/응답 요지·주요 에러코드를 고정한다. 본 문서는 api-surface.md(Control-Plane
    API Surface v1)를 1:1 변환한 D1 codegen 산출물이다.


    요지 수준 규약: components.schemas는 요지(shape) 수준이며, 상세 타입(IR/Verify/Event payload·필드
    required·examples)은 계약 ts/schema 및 codegen/types를 참조한다. 어휘 체인 abort → cancelled →
    run.cancelled, "조용한 false/unknown 금지", error-catalog 메타(httpStatus/exceptionClass)를 그대로 따른다.


    - 인증: 모든 엔드포인트 인증 필수. tenant_id는 요청 본문/쿼리가 아니라 인증 주체(토큰)에서 도출한다(api-surface §0.1).

    - 에러 본문: 모든 4xx/5xx는 ApiError(ts/error-catalog.ts). HTTP 상태코드는 ERROR_CATALOG[code].httpStatus를 그대로 사용한다(api-surface §0.2).

    - DEAD_LETTER(httpStatus 200)는 API 오류 응답이 아니다 — 상태 통지 전용 코드이므로 ApiError로 반환하지 않는다(api-surface §0.2/§4).
servers:
  - url: /v1
    description: Control-plane v1 base path.
...
security:
  - bearerAuth: []
paths:
...
  /runs:
    post:
      tags: [Runs]
      operationId: createRun
      summary: 실행 생성
      description: >
        run 생성. 서버가 params.as_of를 1회 고정한다(요청 명시 시 그 값, 미지정 시 생성 시각;
        api-surface §0.6) → runs.as_of 영속화 → IREL date_* 결정론 보장. run.created 이벤트 emit.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunCreateRequest'
      responses:
        '201':
          description: 생성됨. run 리소스(status=queued). run.created 이벤트 emit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunCommandResponse'
        '403':
          $ref: '#/components/responses/Error'        # SITE_PROFILE_BLOCKED
        '422':
          $ref: '#/components/responses/Error'        # IR_SCHEMA_INVALID / IR_EXPRESSION_COMPILE_ERROR
        default:
          $ref: '#/components/responses/Error'
    get:
      tags: [Runs]
      operationId: listRuns
      description: List tenant-scoped run records for operations monitoring and trace entry points.
      summary: 실행 목록
      parameters:
        - name: status
          in: query
          description: RunState 값으로 필터(state-machine-types.ts RunState).
          required: false
          schema:
            $ref: '#/components/schemas/RunState'
        - name: scenario_version_id
          in: query
          required: false
          schema: { type: string }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: 커서 페이지(items=Run).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunPage'
        default:
          $ref: '#/components/responses/Error'
  /runs/{run_id}:
    parameters:
      - $ref: '#/components/parameters/RunId'
    get:
      tags: [Runs]
      operationId: getRun
      description: Return one run record with current state, worker metadata, and failure context.
      summary: 실행 상세
      responses:
        '200':
          description: run 상세(status ∈ RunState, worker_id, attempts, as_of, current_node, failure_reason).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '404':
          $ref: '#/components/responses/Error'        # RUN_NOT_FOUND
        default:
          $ref: '#/components/responses/Error'
...
  /runs/{run_id}/abort:
    parameters:
      - $ref: '#/components/parameters/RunId'
    post:
      tags: [Runs]
      operationId: abortRun
      summary: 실행 중단(취소) 요청
      description: >
        어휘 정합: 명령은 abort → 상태는 aborting→cancelled(R6/R10/R16/R23/R24/R26/R27/R28) →
        이벤트는 run.cancelled → UI 문구 "취소됨". 엔드포인트명은 abort 유지(api-surface §1).
        completing은 finalize 우선(R25) — abort 거부, RUN_ALREADY_TERMINAL(409). queued/claimed는
        dispatcher 큐/claim 회수로 처리하되 API는 동일하게 202 수락(결과 cancelled로 수렴).
        suspending은 bookmark-cancel port/durable abort intent가 없으면 bookmark 저장 중 side effect를
        추정하지 않고 멱등 예약 전에 WORKITEM_CHECKOUT_CONFLICT(run_bookmark_in_progress)로 fail-closed한다.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AbortRequest'
      responses:
        '202':
          description: abort 수락(aborting 경유 cancelled). run.cancelled 이벤트.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunCommandResponse'
        '404':
          $ref: '#/components/responses/Error'        # RUN_NOT_FOUND
        '409':
          $ref: '#/components/responses/Error'        # RUN_ALREADY_TERMINAL / RUN_ABORTED / WORKITEM_CHECKOUT_CONFLICT(suspending)
        default:
          $ref: '#/components/responses/Error'

  # ============================================================
  # 1.5 Run Triggers (예약 실행 / orchestration)
  # ============================================================
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT bearer. tenant_id 및 roles(Role[]) 클레임에서 도출(auth-rbac §3). 클라이언트 body/query의
        tenant_id는 불신. 인증 미성립(토큰 누락/서명 무효) 시 UNAUTHENTICATED(401), 인증됐으나 tenant
        누락/모호 또는 역할 권한 부족 시 AUTHZ_FORBIDDEN(403). 두 경우 모두 본문은 ApiError.
  # --- Reusable parameters ---
  parameters:
    RunId:
      name: run_id
      in: path
      required: true
      schema: { type: string }
...
    Limit:
      name: limit
      in: query
      required: false
      description: 페이지 크기. 상한 기본값은 운영 정책(Phase 3) — 본 계약은 파라미터 형태만 고정(api-surface §0.5).
      schema:
        type: integer
        minimum: 1
    Cursor:
      name: cursor
      in: query
      required: false
      description: 불투명 커서. 직전 응답 next_cursor 값을 그대로 전달(api-surface §0.5).
      schema:
        type: string
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >
        부작용 명령형 POST의 중복 제출 보호(api-surface §0.4). 서버는 (tenant_id, endpoint, key)로
        최초 결과를 보관, 동일 키 재제출 시 부작용 재실행 없이 최초 응답 반환. sink_idempotency_key(데이터평면)와는 다른 계층.
      schema:
        type: string

  # --- Reusable headers ---
...
  responses:
    Error:
      description: >
        오류 응답. 본문은 ApiError(ts/error-catalog.ts). HTTP 상태는 ERROR_CATALOG[code].httpStatus를
        그대로 사용한다. DEAD_LETTER는 오류 응답이 아니므로 본문 code로 반환되지 않는다(api-surface §0.2).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

  # --- Schemas (요지 수준; 상세 타입은 계약 ts/schema 및 codegen/types 참조) ---
  schemas:
...
    ApiError:
      type: object
      description: 공통 에러 응답(ts/error-catalog.ts ApiError).
      required: [code, message, correlation_id]
      additionalProperties: false
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: userMessage 또는 상세(민감정보 없음).
        details:
          description: 코드별 상세 페이로드(예 IR_SCHEMA_INVALID의 reason). 구조는 D1 위임(api-surface §9).
        correlation_id:
          type: string
          format: uuid
          description: event-envelope·trace correlation_id와 동일 값(api-surface §0.2).
    ErrorCode:
      type: string
      description: >
        ts/error-catalog.ts ErrorCode(48개). DEAD_LETTER는 enum에 포함되나 상태 통지 전용으로
        ApiError 응답 code로는 반환되지 않는다(api-surface §0.2).
      enum:
        - RUN_NOT_FOUND
        - RESOURCE_NOT_FOUND
        - RUN_ALREADY_TERMINAL
        - RUN_ABORTED
        - SCENARIO_VERSION_CONFLICT
        - POLICY_VERSION_CONFLICT
        - IR_SCHEMA_INVALID
        - IR_EXPRESSION_COMPILE_ERROR
        - IR_EXPRESSION_RUNTIME
        - IR_NO_BRANCH_MATCHED
        - SITE_PROFILE_BLOCKED
        - SITE_CIRCUIT_OPEN
        - SESSION_LOCKED
        - SESSION_GENERATION_CONFLICT
        - SESSION_REGISTRATION_REQUIRED
        - CHALLENGE_UNRESOLVED
        - RATE_BUDGET_EXCEEDED
        - BROWSER_LEASE_EXPIRED
        - BROWSER_CRASH
        - CDP_DISCONNECTED
        - NAVIGATION_TIMEOUT
        - LLM_BUDGET_EXCEEDED
        - LLM_CAPABILITY_MISMATCH
        - LLM_STREAM_TIMEOUT
        - LLM_STREAM_IDLE_TIMEOUT
        - LLM_MALFORMED_OUTPUT
        - LLM_CONTENT_FILTERED
        - LLM_RATE_LIMITED
        - LLM_BACKEND_UNAVAILABLE
        - LLM_CONNECTION_FAILED
        - EXTRACT_SCHEMA_INVALID
        - VERIFY_FAILED
        - EMPTY_RESULT_NO_WITNESS
        - SECRET_ACCESS_DENIED
        - DOMAIN_POLICY_VIOLATION
        - PROMPT_INJECTION_DETECTED
        - ARTIFACT_NOT_REDACTED
        - SHELL_COMMAND_NOT_ALLOWED
        - UNAUTHENTICATED
        - AUTHZ_FORBIDDEN
        - CONNECTOR_PERMISSION_DENIED
        - CONNECTOR_INCOMPATIBLE
        - CONNECTOR_HOOK_FAILED
        - SINK_DELIVERY_FAILED
        - RAW_PERSIST_FAILED
        - CONTROL_PLANE_INTERNAL_ERROR
        - HUMAN_TASK_EXPIRED
        - APPROVAL_ALREADY_DECIDED
        - WORKITEM_CHECKOUT_CONFLICT
        - DEAD_LETTER

    # ===== State enums (ts/state-machine-types.ts — 정확히 일치) =====
    RunState:
      type: string
      enum:
        - queued
        - claimed
        - running
        - suspending
        - suspended
        - resume_requested
        - resuming
        - completing
        - completed
        - aborting
        - cancelled
        - failed_business
        - failed_system
...
    RunPage:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Run' }
        next_cursor:
          type: [string, "null"]
...
    Run:
      type: object
      description: 실행 리소스(요지). Read responses always include current_node and failure_reason; unknown node and absent failure are explicit nulls.
      required:
        - run_id
        - scenario_id
        - scenario_version_id
        - status
        - worker_id
        - attempts
        - as_of
        - current_node
        - failure_reason
        - updated_at
      properties:
        run_id: { type: string }
        scenario_id: { type: string }
        status: { $ref: '#/components/schemas/RunState' }
        scenario_version_id: { type: string }
        worker_id: { type: [string, "null"] }
        attempts: { type: integer }
        as_of:
          type: [string, "null"]
          format: date-time
          description: 서버가 생성 시 1회 고정(api-surface §0.6). 재시도/replay/resume에 재사용.
        workitem_id: { type: [string, "null"] }
        current_node:
          type: [string, "null"]
          description: Current executing IR node id when known; null when the runtime has no authoritative node.
        failure_reason:
          type: [object, "null"]
          additionalProperties: false
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
        updated_at:
          type: string
          format: date-time
          description: runs.updated_at. 콘솔이 오래 멈춘 실행을 추정하지 않고 실제 상태 갱신 시각으로 판단하는 데 사용.
    RunCommandResponse:
      type: object
      description: Minimal acknowledgement for run side-effect commands. Read shape is components.schemas.Run.
      additionalProperties: false
      required: [run_id, status]
      properties:
        run_id: { type: string }
        status: { $ref: '#/components/schemas/RunState' }
        as_of:
          type: [string, "null"]
          format: date-time
          description: Present on createRun responses; omitted on abort responses.
...
    RunCreateRequest:
      type: object
      required: [scenario_version_id, params]
      properties:
        scenario_version_id: { type: string }
        params:
          type: object
          description: params_schema 준수(scenario의 params_schema). params.as_of 포함 가능.
        workitem_id:
          type: string
          description: optional. 작업항목 연계 시.
        model:
          type: string
          description: optional. gateway_policies.model override resolved and frozen on run create.
      additionalProperties: false
    AbortRequest:
      type: object
      properties:
        reason: { type: string }
      additionalProperties: false
...